feat: 🎸 Renamed File [LOGOTYPE UTIL]
