+++ frontend/src/components/Utils/LogotypeUtil/index.test.tsx
